pyproject.toml
[project]
name = "scoopwhoop-postgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.58.2",
    "dotenv>=0.9.9",
    "numpy>=2.3.1",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "pillow>=10.0.0",
    "pymongo>=4.6.1",
    "selenium>=4.34.2",
    "streamlit>=1.28.0",
]
